+++ src/index.js
@@ -9,6 +9,7 @@ import { BrowserRouter } from 'react-router-dom';
 import { HelmetProvider } from 'react-helmet-async';
 import { ThemeProvider } from '@mui/material/styles';
 import { theme } from './styles/GlobalStyle';
+import AuthControl from './components/common/AuthControl';
 import App from './App';
 
 const queryClient = new QueryClient({
@@ -28,6 +29,7 @@ root.render(
           <BrowserRouter>
             <HelmetProvider>
               <ThemeProvider theme={theme}>
+                <AuthControl />
                 <App />
               </ThemeProvider>
             </HelmetProvider>
